first attempt to allow pip install
