+++ redcap_importer/templates/redcap_importer/instrument_details.html
@@ -1,0 +1,95 @@
+{% extends "core/base.html" %}
+
+{% block content %}
+
+	<nav aria-label="breadcrumb">
+ 		<ol class="breadcrumb">
+    		<li class="breadcrumb-item"><a href="{% url 'redcap_importer:home' %}">REDCap Connection List</a></li>
+    		<li class="breadcrumb-item">
+    			<a href="{% url 'redcap_importer:project_details_instruments' data_source oProject.id %}">
+    				{{ oProject }}
+    			</a>
+    		</li>
+    		<li class="breadcrumb-item active" aria-current="page">{{ oInstrument }}</li>
+  		</ol>
+	</nav>
+
+	<h1>{{ oInstrument }}</h1>
+	
+	<hr>
+	<h2>Instrument Details</h2>
+	<table class="table table-dark">
+		<thead>
+			<tr>
+				<th>Instrument</th>
+				<th>Label</th>
+				<th>Project</th>
+				<th>Record Count</th>
+			</tr>
+		</thead>
+		<tbody>
+				<tr>
+					<td>{{ oInstrument.unique_name }}</td>
+					<td>{{ oInstrument.label }}</td>
+					<td>{{ oProject.project_title }}</td>
+					<td>{{ oInstrument.get_count }}</td>
+				</tr>
+		</tbody>
+	</table>
+	
+	
+	<h2>Fields</h2>
+	<table class="table table-bordered table-striped table-sm">
+		<thead>
+			<tr>
+				<th>Order</th>
+				<th>Field</th>
+				<th>Label</th>
+				<th>Field Type</th>
+				<th>Select Type</th>
+				<th class="text-primary">Count</th>
+				<th class="text-success">Count Non-Null</th>
+				<th class="text-danger">Count Null</th>
+				<th>Count Distinct</th>
+				<th>Min</th>
+				<th>Max</th>
+			</tr>
+		</thead>
+		<tbody>
+			{% for oField in oInstrument.fieldmetadata_set.all %}
+				{% with stats=oField.get_stats.0 %}
+					<tr>
+						<td>{{ oField.ordering }}</td>
+						<th>
+							<a href="{% url 'redcap_importer:field_details' data_source oProject.id oField.id %}">
+								{{ oField.unique_name }}
+							</a>	
+						</th>
+						<td>{{ oField.label|truncatechars:75 }}</td>
+						<td>{{ oField.django_data_type }}</td>
+						<td>
+							{% if oField.is_many_to_many %}
+								MultiSelect
+							{% elif oField.get_display_lookup %}
+								Select
+							{% endif %}
+						</td>
+						<td class="text-primary">{{ stats.count }}</td>
+						<td class="text-success">{{ stats.count_not_null }}</td>
+						<td class="text-danger">{{ stats.count_null }}</td>
+						<td>{{ stats.distinct }}</td>
+						<td>{{ stats.min|truncatechars:50 }}</td>
+						<td>{{ stats.max|truncatechars:50 }}</td>
+					</tr>
+				{% endwith %}
+			{% endfor %}
+		</tbody>
+	</table>
+	
+	
+	
+	
+	
+{% endblock %}
+
+
