--- templates/redcap_importer/project_details_events.html
@@ -1,21 +1,0 @@
-
-{% extends "redcap_importer/containers/project_details_container.html" %}
-
-{% block content2 %}	
-	
-	{% if oProject.is_longitudinal %}
-		<h2>Events</h2>
-		<div class="card-columns">
-			{% for oEvent in oProject.eventmetadata_set.all %}
-				{% include 'redcap_importer/snippets/event_card.html' %}
-			{% endfor %}
-		</div>
-	{% endif %}
-	
-	
-	
-{% endblock %}	
-	
-
-
-
