--- templates/redcap_importer/project_details_events_multi_arm.html
@@ -1,43 +1,0 @@
-
-{% extends "redcap_importer/containers/project_details_container.html" %}
-
-{% block content2 %}	
-
-	{% if oProject.is_longitudinal %}
-		<h2>Arms/Events</h2>
-
-		<ul class="nav nav-pills mt-3">
-			{% for arm in oProject.get_events_grouped_by_arm %}
-				<li class="nav-item">
-	    			<a class="nav-link {% if forloop.first %}active{% endif %}" data-toggle="tab" href="#arm-tab-{{arm}}">
-	    				ARM {{ arm }}
-	    			</a>
-	  			</li>
-	  		{% endfor %}
-		</ul>
-
-
-		<div class="tab-content mt-3">
-			{% for arm, qEvent in oProject.get_events_grouped_by_arm.items %}
-	  			<div class="tab-pane {% if forloop.first %}active{% else %}fade{% endif %}" id="arm-tab-{{arm}}">
-	  				<div class="card-columns">
-		  				{% for oEvent in qEvent %}
-							{% include 'redcap_importer/snippets/event_card.html' %}
-						{% endfor %}
-					</div>
-	  			</div>
-  			{% endfor %}
-		</div>
-
-	{% endif %}
-
-
-
-{% endblock %}	
-	
-	
-
-	
-
-
-
